fix(clippy): fix unused import, expected client type, and unnecessary cast
Fixes Clippy and compiler errors in scout_access and verification contracts: imports of FeeConfig re-export, ProgressContractClient struct initialization, and unnecessary entries.len() as u32 cast.

+++ contracts/verification/src/lib.rs
@@ -386,7 +386,7 @@ impl VerificationContract {
             .instance()
             .get(&DataKey::ActiveValidatorCount)
             .unwrap_or(0u32);
-        let batch_len = entries.len() as u32;
+        let batch_len = entries.len();
         if current_count
             .checked_add(batch_len)
             .ok_or(VerificationError::Overflow)?
